training_pipeline/requirements.txt
numpy==2.0.2
torch==2.8.0
loguru==0.7.3
tqdm==4.67.1
mlflow==3.4.0
lightning==2.5.5
pandas==2.3.2
evidently==0.6.5
pyarrow==17.0.0
pydantic==2.10.6
gcsfs==2024.6.1
google-cloud-storage==2.19.0
scikit-learn==1.7.2
dill==0.3.8
torchmetrics==1.8.2
qdrant-client==1.12.1
feast[gcp,redis]==0.50.0
scipy==1.16.2
redis==2.4.0